requirements.txt
aiohttp==3.9.5
httpx==0.27.0
numpy<2.0.0,>=1.0
openai<2.0.0,>=1.32.0
regex==2024.5.15
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.43.1
tqdm==4.66.4
python-docx==1.1.2
markdownify==1.0.0
sentence_transformers==3.4.1
faiss-cpu==1.10.0
rapidfuzz==3.12.1
spacy==3.8.4
html2text==2024.2.26
matplotlib==3.10.1
seaborn==0.13.2
markdown==3.7
